fixing the su issue and created the super function

+++ bootstrap.sh
@@ -10,27 +10,48 @@ if [ -z "${BOOTSTRAP_UP_TO_DATE}" ]; then
 fi
 
 # start of the actual script
+DISTRO=$(grep -E '^NAME=' /etc/os-release | cut -d= -f2 | tr -d '"')
 SCRIPT_PATH="$( cd -- "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
 
-set -e
+###############################################################################
+# Command setup section
+###############################################################################
+if command -v rsync > /dev/null; then
+	CMD_COPY="rsync -av"
+else
+	CMD_COPY="cp -r"
+fi
 
-# creating a temporary password file
-PASSWORD_FILE=$(mktemp)
-trap 'rm -f "$PASSWORD_FILE"' EXIT
-chmod 600 "$PASSWORD_FILE"
+super() {
+	if command -v sudo > /dev/null; then
+		sudo "$1"
+	else
+		if [ ! -e "${PASSWORD_FILE}" ]; then
+			# Password handling should never fail
+			set -e
 
-printf "Password: "
-stty -echo
-read -r PASSWORD
-stty echo
-echo ""
+			# creating a temporary password file
+			PASSWORD_FILE=$(mktemp)
+			trap 'rm -f "$PASSWORD_FILE"' EXIT
+			chmod 600 "$PASSWORD_FILE"
 
-# storing the password
-echo "$PASSWORD" > "$PASSWORD_FILE"
-unset PASSWORD
+			printf "Password: "
+			stty -echo
+			read -r PASSWORD
+			stty echo
+			echo ""
 
-set +e
+			# storing the password
+			echo "$PASSWORD" > "$PASSWORD_FILE"
+			unset PASSWORD
 
+			# we can allow failure now
+			set +e
+		fi
+
+		su -c "$1" root < "${PASSWORD_FILE}"
+	fi
+}
 ###############################################################################
 # Windows Detection
 ###############################################################################
@@ -63,12 +84,7 @@ if $WINDOWS; then
 	ln -sf "/mnt/c/Users/${WINDOWS_USER}/Pictures"			"${HOME}/Pictures"
 	ln -sf "/mnt/c/Users/${WINDOWS_USER}/Videos"				"${HOME}/Videos"
 
-	if command -v rsync > /dev/null; then
-		rsync -av "${SCRIPT_PATH}/windows/etc" "/etc/"
-	else
-		cat "${PASSWORD_FILE}" | su -c "cp -r \"${SCRIPT_PATH}/windows/etc\" /etc/" root
-	fi
-
+	super -c "${CMD_COPY} \"${SCRIPT_PATH}/platform/windows/etc\" /etc/" root < "${PASSWORD_FILE}"
 else
 	echo "This is not a Windows (WSL) environment. Skipping Windows-specific configurations."
 
@@ -86,19 +102,16 @@ fi
 ###############################################################################
 #   Syncing configuration files
 ###############################################################################
+${CMD_COPY} "${SCRIPT_PATH}/home/"		"${HOME}/"
+${CMD_COPY} "${SCRIPT_PATH}/config/"	"${HOME}/.config/"
 
-if command -v rsync > /dev/null; then
-	rsync -av "${SCRIPT_PATH}/home/"		"${HOME}/"
-	rsync -av "${SCRIPT_PATH}/config/"	"${HOME}/.config/"
+super -c "${CMD_COPY} \"${SCRIPT_PATH}/etc/\"		/etc/" root < "${PASSWORD_FILE}"
 
-	cat "${PASSWORD_FILE}" | su -c "rsync -av \"${SCRIPT_PATH}/etc/\"		/etc/" root
-else
-	cp -r "${SCRIPT_PATH}/home/"		"${HOME}/"
-	cp -r "${SCRIPT_PATH}/config/"	"${HOME}/.config/"
-
-	echo "# Copying /etc configurations"
-	cat "${PASSWORD_FILE}" | su -c "cp -r \"${SCRIPT_PATH}/etc/\"		/etc/" root
-fi
+case "$DISTRO" in
+	"Arch Linux")
+		super -c "${CMD_COPY} \"${SCRIPT_PATH}/platform/arch/etc/\"		/etc/" root	< "${PASSWORD_FILE}"
+		;;
+esac
 
 ###############################################################################
 #   SSH KEY CREATION
@@ -111,5 +124,10 @@ if [ ! -f "${HOME}/.ssh/device_ssh_key" ]; then
 	echo "[ ssh-keygen ] creating ~/.ssh/device_ssh_key"
 	ssh-keygen -t ed25519 -f "${HOME}/.ssh/device_ssh_key" -N ""
 fi
-	
+
+###############################################################################
+#   Storing distribution name in /etc/environment
+###############################################################################
+super -c "echo 'DISTRO=${DISTRO}' | tee -a /etc/environment > /dev/null" < "${PASSWORD_FILE}"
+
 unset BOOTSTRAP_UP_TO_DATE
